requirements.txt
# Google Cloud libraries
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform

# Data processing and analysis
pandas
numpy
datetime

# API requests
requests

# Data visualization
matplotlib
seaborn
plotly

# Jupyter notebooks (optional, for interactive development)
jupyter

# Machine learning
scikit-learn
tensorflow

# Geospatial data handling (for GPX files)
gpxpy

# Web framework (if needed for deployment)
flask

# Environment management
python-dotenv

# Interactive widgets
ipywidgets
dash